Game: Added more game logic around hypothesises.
Player:
Added board to player constructor to complete association.
Renamed some poorly named fields.
Created a method for displaying hand as text.
Created a method to check if player allowed to act.
Created a method to check if player in a room.

Position:
Added .equals and hashCode

Tile:
Refactored for clarity.